+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - TT-1795 - Display error message to second user in registation journey
 - TT-1841 - Email not displayed on invite new admin
 - TT-1821 - Change `I can't find my business` link on company's house registration.
+- XOT-1120 - Take user back to the ingress url when registering
 - TT-1786 - Change 'start now' text on wizard
 - no ticket - Replace `DIRECTORY_CONSTANTS_URL_GREAT_INTERNATIONAL` with `DIRECTORY_CONSTANTS_URL_INTERNATIONAL`
 - TT-1903 - Removed Editor from change role in admin panel
@@ -64,6 +65,8 @@
 - TT-1901 - show name in collaborators list 
 - TT-1914 - Reinstate missing terms tickbox
 - TT-1915 - Fix update profile error if user has no supplier
+- TT-1902 - Collaborator Invite UI bug fixes
+- TT-1946 - Remove 'downgrade\upgrade to editor' options in admin tools
 
 ## [2019.08.14_1](https://github.com/uktrade/directory-ui-supplier/releases/tag/2019.08.14_1)
 [Full Changelog](https://github.com/uktrade/directory-ui-supplier/compare/2019.08.14...2019.08.14_1)